+++ core/consts.c
@@ -1,0 +1,13 @@
+const int FPS = 30;
+const int DISPLAY_WIDTH = 1280;
+const int DISPLAY_HEIGHT = 720;
+const int MAP_SIZE = 2336;
+const int INITIAL_MAP_X = -320;
+const int INITIAL_MAP_Y = -90;
+const int PLAYER_WIDTH = 100;
+const int PLAYER_HEIGHT = 125;
+const int PLAYER_SPEED = 5;
+const int BUTTON_WIDTH = 200;
+const int BUTTON_HEIGHT = 50;
+const int HEART_WIDTH = 220;
+const int HUNGER_WIDTH = 301;
